Use `pentest-chef-repo` to provision kali
This repository located at
https://github.com/Sliim/pentest-chef-repo.git provides some roles that
can be used directly to setup some tools. For now, it provides roles for
base and full kali install, proxy configuration with tor and proxychains,
and a Faraday setup.

To install `pentest-chef-repo`, simply init and update git
submodules. If they are missing, kali linux will not be provisioned with
Chef.